[FLINK-11368][tests] Port TaskManagerStartupTest to new code base
